Add fixed scale enum (not really supported though)

+++ network-area-diagram-viewer/src/main/java/com/powsybl/nad/viewer/MainViewController.java
@@ -87,7 +87,7 @@ public class MainViewController implements ChangeListener<Object> {
     @FXML
     public ChoiceBox<SvgParameters.SizeConstraint> sizeConstraintChoice;
     @FXML
-    public Spinner<Integer> fixedSizeSpinner;
+    public Spinner<Double> fixedSizeSpinner;
 
     @FXML
     public TabPane checkedTab;
